Extract render-search-results-list
